# When parsing, validate that chars are in expected Base32 range.