This adds new validation in `UnmarshalText` and `Parse` that returns an `ErrInvalidCharacters` error when called with input bytes outside the expected Base32 encoding set. See https://github.com/oklog/ulid/issues/21 for details.

This does unfortunately end up being a bit slower. Some basic benchmarks from my laptop:

#### Before
```
BenchmarkUnmarshal/Text-8         	1000000000	        17.3 ns/op	1506.08 MB/s
```

#### With a simple loop
```
BenchmarkUnmarshal/Text-8         	500000000	        32.5 ns/op	 799.66 MB/s
```

#### With a manually unrolled loop (this change)
```
BenchmarkUnmarshal/Text-8         	500000000	        26.6 ns/op	 976.36 MB/s
```

#### Comparison from original to this version
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkUnmarshal/Text-8       17.3          26.6          +53.76%

benchmark                       old MB/s     new MB/s     speedup
BenchmarkUnmarshal/Text-8       1506.08      976.36       0.65x
```

Fixes https://github.com/oklog/ulid/issues/21.