@optiomas 2018-01-30
I’d make this a separate and optional function: `IsValid`