@optiomas 2018-01-30
Defining ParseStrict is trivial on the client side, given the availability of IsValid. Opinions @peterbourgon?