@peterbourgon 2018-01-30
Having Parse and ParseStrict seems like a nicer API to me.