@optiomas 2018-01-30
Would there be possible optimizations to try out before we go for ParseStrict? If we can’t get the code to a similar performance range, then let’s do it.