@mattmoyer 2018-01-31
I couldn't think of any other optimizations to try here. I did try the guard as a series of returns rather than a big `||` chain, which generated different code but with pretty much identical performance:
```go
if dec[v[0]] == 0xFF {
    return ErrInvalidCharacters
}
if dec[v[1]] == 0xFF {
    return ErrInvalidCharacters
}
# [...]
```

I also tried using `regexp`, thinking maybe there were some optimized routines there:
```go
var ValidULIDPattern = regexp.MustCompile(`\A[0123456789ABCDEFGHJKMNPQRSTVWXYZabcdefghjkmnpqrstvwxyz]{26}\z`)
#[...]
```

It did not work very well:
```
BenchmarkUnmarshal/Text-8       17.3          586           +3287.28%
```


Is parsing text ULIDs performance critical for most use cases? I'd expect ULID generation and `.String()` to be sensitive use cases but maybe not so much `Parse()`.