@optiomas 2018-01-31
> Is parsing text ULIDs performance critical for most use cases? I'd expect ULID generation and .String() to be sensitive use cases but maybe not so much Parse().

It may be, but I'm cool with having the this cost incurred in the new `ParseStrict` only.