@nickform 2018-02-01
Just hit this today. I think it would be best to add the checks into the existing methods as is done in this pull request. My reasoning is PoLS: I fully expected this to be one of the possible error cases already.

My use case for Parse is input validation where correctness is more important than out-and-out speed and I share the doubt raised by @mattmoyer about whether performance here is a reason enough to complicate the API.