@optiomas 2018-02-01
> My use case for Parse is input validation where correctness is more important than out-and-out speed and I share the doubt raised by @mattmoyer about whether performance here is a reason enough to complicate the API.

The driving force behind how this package was designed and implemented was performance and configurability. I won't introduce a performance regression in `Parse` because of this expectation. `ParseStrict` is a good compromise.