@nickform 2018-02-01
Certainly the extra API complexity is no disaster - it's for the package maintainers to make the trade-off they think is best.

While I'm writing - thanks to all the maintainers of this package; it has come in very handy for us.