@mattmoyer 2018-02-02
I'm happy to split out a `ParseStrict` that adds new validation. I'll update this PR shortly.

Another option for discussion:
 - `Parse` gets this validation logic and becomes safe by default.
 - We add a new method like `ParseFast`/`ParseUnsafe`/`ParseQuick` that keeps the performance of the current `Parse` but allows otherwise "invalid" ULIDs to unmarshal.