@optiomas 2018-02-03
> Parse gets this validation logic and becomes safe by default.

Do you see the current behaviour as unsafe? Granted, it's unexpected, but I'm failing to see how it's unsafe.

> We add a new method like ParseFast/ParseUnsafe/ParseQuick that keeps the performance of the current Parse but allows otherwise "invalid" ULIDs to unmarshal.

I'd like to prevent the introduction of 1) performance regressions and 2) unexpected errors by users of this package. `ParseStrict` solves that.