@nickform 2018-02-03
What about an unbound ```isValid(string)``` method which you can optionally call before parsing? This would be sufficient for input validation - we just discard the result of parsing anyway.