@optiomas 2018-02-03
That was my first suggestion @nickform. After all this conversation I feel that’s indeed better.