@nickform 2018-02-03
My apologies - I had interpreted your original suggestion as a method bound to the ULID type but I see that you said function.