@tsenart 2018-07-12
@peterbourgon: Revisiting this. Thinking of the user base of this library, I'm not so sure anymore if the 80% would feel the performance regression introduced by this change set. On the other hand, 100% of it would benefit from better validation immediately. Do you think there are 20% of users that rely heavily on the performance of the unmarshalling code as it stands?