@peterbourgon 2018-07-12
@tsenart It's a judgment call and I'm not sure. Therefore, I'm most in favor of leaving `Parse` alone and adding new `ParseStrict` with new validation behavior.