@tsenart 2018-07-12
@peterbourgon: Although definitely not as significant, I'm afraid we weren't as diligent with performance here: https://github.com/oklog/ulid/pull/20. Would have been good to capture there as a point of comparison.