@tsenart 2018-07-12
https://github.com/RobThree/NUlid/blob/master/NUlid/Ulid.cs#L262 tells me there is another way to validate the input. I'd like to compare the performance of that approach before settling on the API changes.