@mattmoyer 2018-07-13
Hey all, sorry for the long delay here.

I rebased this PR, but also opened https://github.com/oklog/ulid/pull/26 with the `ParseStrict()`/`UnmarshalTextStrict()` approach.

@tsenart @peterbourgon feel free to pick which approach you're most comfortable with and merge/close as appropriate. Either one is fine by me.